Remove duplicate verapdf TestCase and autoload shim
Closes mooxphp/verapdf#4
